Added abstarct class to avoid code copying!
